requirements.txt
beautifulsoup4==4.6.3
cffi==1.11.5
Cython==0.28.5
Django==2.1.1
easydict==1.8
editdistance==0.5.2
gevent==1.2.2
GPUtil==1.3.0
gunicorn==19.9.0
htmlmin==0.1.12
lmdb==0.94
msgpack==0.5.6
nltk==3.3
numpy==1.15.1
opencv-python==3.4.3.18
Pillow==5.3.0
pyexcel-xls==0.5.7
pymongo==3.7.2
PyMuPDF==1.14.2
PyMySQL==0.9.3
pytest==3.3.2
python-box==3.2.0
python-docx==0.8.6
python-magic==0.4.15
PyYAML==3.13
redis==3.0.1
requests==2.21.0
rpyc==4.0.2
scipy==1.1.0
Shapely==1.6.4.post2
SQLAlchemy==1.2.1
torchvision==0.2.1
tqdm==4.26.0
unidecode==1.0.22
